update:  refactor: v1.0.0完成としてマーク, ファイル整理

+++ Project/sources/00_Engine/00-1_Core/Utility.cpp
@@ -118,7 +118,7 @@ float Utility::Easing::CalculateEase(double current, double duration, easing_fun
 {
 	double ratio = CalculateRatio(current, duration);
 
-	float ease = getEasingFunction(easeType)(ratio);
+	float ease = static_cast<float>(getEasingFunction(easeType)(ratio));
 
 	return ease;
 }
